pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "SINC-rPPG"
version = "0.1.0"
description = "Non-Contrastive Unsupervised Learning of Physiological Signals from Video"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "h5py>=3.12.0",
    "matplotlib>=3.9.0",
    "mediapipe>=0.10.11",
    "natsort>=8.4.0",
    "numpy>=1.26.0,<2.0.0",
    "opencv-contrib-python>=4.9.0.80",
    "pandas>=2.2.0",
    "scipy>=1.13.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "tqdm>=4.66.1",
    "tensorboard>=2.20.0",
    "lightning>=2.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[tool.setuptools]
package-dir = { "" = "src" }
py-modules = ["args", "train", "test", "repo_paths", "config_merge"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
